refactor: move user-fetch process and form control to parent component
